[WFCORE-4956] Move retrieving and setting the TCCL from the public API to a package protected API.
