+++ embedded/src/main/java/org/wildfly/core/embedded/EmbeddedHostControllerFactory.java
@@ -253,9 +253,9 @@ public class EmbeddedHostControllerFactory {
 
         @Override
         public void start() throws EmbeddedProcessStartException {
-            ClassLoader tccl = EmbeddedManagedProcess.getTccl();
+            ClassLoader tccl = SecurityActions.getTccl();
             try {
-                EmbeddedManagedProcess.setTccl(embeddedModuleCL);
+                SecurityActions.setTccl(embeddedModuleCL);
                 EmbeddedHostControllerBootstrap hostControllerBootstrap = null;
                 try {
                     final long startTime = System.currentTimeMillis();
@@ -294,7 +294,7 @@ public class EmbeddedHostControllerFactory {
                     throw EmbeddedLogger.ROOT_LOGGER.cannotStartEmbeddedServer(ex);
                 }
             } finally {
-                EmbeddedManagedProcess.setTccl(tccl);
+                SecurityActions.setTccl(tccl);
             }
         }
 
@@ -360,12 +360,12 @@ public class EmbeddedHostControllerFactory {
 
         @Override
         public void stop() {
-            ClassLoader tccl = EmbeddedManagedProcess.getTccl();
+            ClassLoader tccl = SecurityActions.getTccl();
             try {
-                EmbeddedManagedProcess.setTccl(embeddedModuleCL);
+                SecurityActions.setTccl(embeddedModuleCL);
                 exit();
             } finally {
-                EmbeddedManagedProcess.setTccl(tccl);
+                SecurityActions.setTccl(tccl);
             }
         }
